requirements.txt
# Core
playwright==1.48.0
pytest==8.3.4
pytest-playwright==0.5.2
pytest-html==4.1.1
pytest-xdist==3.6.1

# API Testing
requests==2.32.3

# Utilities
python-dotenv==1.0.1
faker==30.8.2

# Reporting
allure-pytest==2.13.5

# Data
pydantic==2.10.3